64-bit
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@1025 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ DCMPix.m
@@ -2708,9 +2708,9 @@ long BresLine(int Ax, int Ay, int Bx, int By,long **xBuffer, long **yBuffer)
 	unsigned char   *argbImage, *tmpPtr, *srcPtr, *srcImage;
 	long			i, x, y, totSize;
 	int				realwidth;
-	long	w, h, row;
-	short   bpp, count, spp;
-	short   cur_page, number_of_pages;
+	int				w, h, row;
+	short			bpp, count, spp;
+	short			cur_page, number_of_pages;
 	
 	isRGB = NO;
